Fix exercise call option with hire
