+++ programs/dexloan-listings/src/processor/call_option/mod.rs
@@ -1,11 +1,9 @@
 pub mod buy;
 pub mod close;
 pub mod exercise;
-pub mod exercise_with_hire;
 pub mod initialize;
 
 pub use buy::*;
 pub use close::*;
 pub use exercise::*;
-pub use exercise_with_hire::*;
 pub use initialize::*;
