+++ programs/dexloan-listings/src/processor/loan/mod.rs
@@ -3,11 +3,9 @@ pub mod initialize;
 pub mod give;
 pub mod repay;
 pub mod repossess;
-pub mod repossess_with_hire;
 
 pub use close::*;
 pub use initialize::*;
 pub use give::*;
 pub use repay::*;
 pub use repossess::*;
-pub use repossess_with_hire::*;
